requirements.txt
matplotlib==3.10.0
munkres==1.1.4
numpy>=2.2.2
pandas
scipy>=1.15.1
seaborn
statsmodels
jupyter_client
jupyter_core
ipython
ipykernel
pyyaml
openpyxl
xlrd
psutil
Pygments
pyparsing
python-dateutil
pytz
tzdata
tornado
traitlets
typing_extensions
wcwidth
